A few hacks, but mostly good stuff for typing trove modules
string-dict now has annotations defined in JS (via type-util), which are
converted to true Types via compile-structs.

The rest of the stuff is a bit of hackery to make some builtins worked out, and
names are still being smoothed over, so more cleanup to come for the places
where "builtin" tests appear.

But importing string-dict now comes with types, defined locally with the
module.