Experimental/hacky support for a stdin-reading REPL
Build with

$ make src/scripts/run-interactions.jarr # note the j

Then you can run

$ node src/scripts/run-interactions.jarr

The process waits for stdin up until a line that says "#go" and then runs the
preceding lines as an interaction, stopping on #done. So a program like this
could drive it programmatically:

```
import random

for i in range(0, 100):
  coef = random.randint(5, 25)
  program = """
fun f%s(x): x * %s end
check:
  f%s(%s) is %s
  f%s(4) is 3
end
""" % (i, coef, i, 7, coef * 7, i)

  print(program)

print("#done")
```

And then use a command like:

$ python3 test-input.py | node src/scripts/run-interactions.jarr > out.txt

To drive a bunch of generated programs through. By starting the node process
separately a Python driver could use the results on stdout to compute the next
program to provide as input.

This still takes over 100ms per program, which is better but not great, and
there might be some opportunities internal to the compiler/runtime to speed it
up further.